requirements.txt
PyQt6==6.5.0
Pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
pytesseract==0.3.8